[update] live olo

+++ app/Http/Controllers/CNOP/Transaksi/DualHomingController.php
@@ -687,26 +687,39 @@ class DualHomingController extends Controller
         DB::beginTransaction();
         try {
             
-            TrWoSite::where('ba_id', $id)->where('tipe_ba', 'DUAL_HOMING')
-            ->update(array(
-                'ba_id' => null,
-            )); 
+            $check = TrBa::where('id', $id)->first();
+            
+            if ($check) {
+                $no_dokumen = $check->no_dokumen;
 
-            TrBa::where('id', $id)->where('tipe', 'DUAL_HOMING')->delete();
-        
-            $path = storage_path().'/app/public/pdf/'.$id .'.pdf';
-
-            if(file_exists($path))
-                unlink($path);
-
-
-            DB::commit();
+                TrWoSite::where('ba_id', $id)->where('tipe_ba', 'DUAL_HOMING')
+                ->update(array(
+                    'ba_id' => null,
+                )); 
     
-            return response()->json([
-                'status' => true,
-                'message' => 'success',
-                'data' => null
-            ], 200);
+                TrBa::where('id', $id)->where('tipe', 'DUAL_HOMING')->delete();
+    
+                $path = storage_path().'/app/public/pdf/'.$id .'.pdf';
+    
+                if(file_exists($path))
+                    unlink($path);
+    
+                MaNomorDokumen::where('no_dokumen', $no_dokumen)->delete();
+
+                DB::commit();
+        
+                return response()->json([
+                    'status' => true,
+                    'message' => 'success',
+                    'data' => null
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'Data Tidak Ditemukan',
+                    'data' => null
+                ], 400);
+            }
 
         } catch (\Exception $e) {
             DB::rollback();
